Rename FlipSet to VEdgeSet and move it to the VEdge class
The reason for this is that we will soon be using this to describe undo
action list, and for that we care about a set of virtual edges without
directions. This adds another use for that structure, so it shouldn't
be called a FlipSet if we are using it in other places.